requirements.txt
pynput==1.7.6
numpy==1.22.4
opencv-python==4.5.5.64
pillow==9.1.1
pygit2==1.9.2
# pip install opencv-contrib-python
picamera[array]==1.13